Update about.html
Removed CSS class "link" from topnav links to differentiate between them and bottom links on index page.

Added a class to the h1 tag, "about" so that I could give it a top margin and "push it down" from the topnav, which is now static.

Fixed a small error with the em tag and a fullstop, and the name of my book.

Added a footer, built with JS, to include a font awesome yin yang. Added font awesome info into the script tag