Update styles.css
Made static topnav using position: fixed.

Put margin:top  on
.about
.contact and
the card to push them down from the  new static topnav bar.

Used Mozilla colour dropper to get a colour from the map and use that for the link animations (hover).

Used rgb to find suitable colour for static topnav.

Separated out the css for the topnav so that there are separate classes for topnav li and topnav li a.

Added a hover pseudo class to the topnav links, giving the animation a colour based on the rotating map, using colour dropper.

Gave the card a border radius, so it would be consistent with the rest of the page.