Update styles.css
I changed the colour of the links so that when active they would show a colour consistent with the webpage, rather than a default.

I removed "position relative" from the body. Not sure why I put that in there but it doesn't seem to do anything.

I removed the keyframes and the "animation duration". I removed the word "transform". I seem to have erroneously treated it as an attribute of the transition property.